+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Features
 
 - Adds tags with Expo Updates context variables to make them searchable and filterable ([#5788](https://github.com/getsentry/sentry-react-native/pull/5788))
+- Automatically capture a warning event when Expo Updates performs an emergency launch ()
 
 ## 8.3.0